Move pickQueryParamsMatchingSchema to root
