+++ src/helpers/index.ts
@@ -2,4 +2,3 @@ export * from './getRawQueryParams';
 export * from './getQueryParamsFromURL';
 export * from './getDefaultQueryParamsState';
 export * from './buildQueryString';
-export * from './pickQueryParamsMatchingSchema';
